fix unit tests of aes_encryptor
